feature/refactor-easy-security
- Refactor EasySecurity
